Feature/test watcher (#6)
* Invalid json handling in Scikit runtime

* Package fixes, Revolver and new config

* Fixes

* Flat repo and fixes

* Tensorflow support

* ModelFiles table and events distinction

* Tensorflow metadata fetching

* Smart indexing

* Create and delete indexing

* - Added test for ServiceAPI

* - Added Jenkinsfile

* - Added SBT
* - Added repository
* Metadata separation and source repo
* More tests
* Test config
* Local source paths fix
* Local watcher tests
* AWS mocks
* Mock revert + cleanup

* Export jar fix

* Merge branch 'master' of https://github.com/Hydrospheredata/hydro-serving into feature/test_watcher

# Conflicts:
#	Jenkinsfile
#	manager/src/main/scala/io/hydrosphere/serving/manager/ManagerServices.scala
#	manager/src/main/scala/io/hydrosphere/serving/manager/service/modelbuild/ModelBuildService.scala
#	project/Dependencies.scala

+++ manager/src/main/scala/io/hydrosphere/serving/manager/service/modelsource/S3ModelSource.scala
@@ -1,69 +1,101 @@
 package io.hydrosphere.serving.manager.service.modelsource
 
-import java.io.File
+import java.io.{File, FileNotFoundException}
 import java.net.URI
 import java.nio.file.{Files, Path, Paths, StandardCopyOption}
 
-import awscala.s3.S3
 import io.hydrosphere.serving.manager.{LocalModelSourceConfiguration, S3ModelSourceConfiguration}
-
-import scala.collection.concurrent.TrieMap
+import org.apache.logging.log4j.scala.Logging
+import scala.collection.JavaConversions._
 
 /**
   * Created by Bulat on 31.05.2017.
   */
-class S3ModelSource(val configuration: S3ModelSourceConfiguration) extends ModelSource {
-  private [this] val localFolder =  s"/tmp/${configuration.name}"
+class S3ModelSource(val configuration: S3ModelSourceConfiguration) extends ModelSource with Logging {
+  private[this] val localFolder =  s"/tmp/${configuration.name}"
   private[this] val fileCache = new LocalModelSource(
     LocalModelSourceConfiguration(s"proxy-${configuration.name}", localFolder)
   )
-  private[this] implicit val s3 = S3.at(configuration.region)
-  private[this] val bucketObj = s3.bucket(configuration.bucket).get
 
-  private def downloadObject(objectPath: String): File = {
-    val fileStream = bucketObj.getObject(objectPath).get.content
-    Files.createDirectories(Paths.get(localFolder, objectPath.split("/").dropRight(1).mkString))
+  val client = configuration.s3Client
+
+  val proxyFolder: Path = Paths.get(localFolder)
+
+  def deleteProxyObject(objectPath: String): Boolean = {
+    Files.deleteIfExists(Paths.get(localFolder, objectPath))
+  }
+
+  def downloadObject(objectPath: String): File = {
+    logger.debug(s"downloadObject: $objectPath")
+    if (!client.doesObjectExist(configuration.bucket, objectPath))
+      throw new FileNotFoundException(objectPath)
+    val fileStream = client.getObject(configuration.bucket, objectPath).getObjectContent
+
+    val folderStructure = Paths.get(localFolder, objectPath.split("/").dropRight(1).mkString("/"))
+
+    Files.createDirectories(folderStructure)
+
     val file = Files.createFile(Paths.get(localFolder, objectPath))
     Files.copy(fileStream, file, StandardCopyOption.REPLACE_EXISTING)
     file.toFile
   }
 
   override def getSubDirs(path: String): List[String] = {
+    logger.debug(s"getSubDirs: $path")
     val pUri = Paths.get(path).toUri
-    val r = bucketObj.keys(path)
+    client
+      .listObjects(configuration.bucket)
+      .getObjectSummaries
+      .map(_.getKey)
       .map(Paths.get(_).toUri)
       .map(k => pUri.relativize(k).toString.split("/").head)
       .distinct.toList
-    r
   }
 
   override def getSubDirs: List[String] = {
-    val r = bucketObj.keys().map(_.split("/").head).distinct.toList
+    logger.debug(s"getSubDirs")
+    val r = client
+      .listObjects(configuration.bucket)
+      .getObjectSummaries
+      .map(_.getKey)
+      .map(_.split("/").head)
+      .distinct
+      .toList
+    println(r)
     r
   }
 
   override def getAllFiles(folder: String): List[String] = {
-    val pUri = URI.create(s"$folder")
-    val modelKeys = bucketObj.keys(pUri.toString).toList
+    logger.debug(s"getAllFiles: $folder")
+    val modelKeys = client
+      .listObjects(configuration.bucket, folder)
+      .getObjectSummaries
+      .map(_.getKey)
+      .filterNot(_.endsWith("/")) // fix w2v/stages/0_w2v_617dd94f64cf/data/
+    println(modelKeys)
     modelKeys
-      .map(k => URI.create(k))
-      .map(k => pUri.relativize(k))
-      .map(k => getReadableFile(k.toString))
+      .map(URI.create)
+      .map(_.toString)
+      .map(getReadableFile)
 
-    fileCache.getAllFiles(folder)
+    val r = fileCache.getAllFiles(folder)
+    println(r)
+    r
   }
 
   override def getReadableFile(path: String): File = {
+    logger.debug(s"getReadableFile: $path")
+
     val fullObjectPath = URI.create(s"$path")
-    println(fullObjectPath)
     val file = fileCache.getReadableFile(path)
     if (!file.exists()) downloadObject(fullObjectPath.toString)
     file
   }
 
-  override def getSourcePrefix(): String = configuration.name
+  override def getSourcePrefix: String = configuration.name
 
-  override def getModelPath(modelSource: String): Path = ???
-
-  override def getLocalCopy(source: String): Path = ???
+  override def getAbsolutePath(modelSource: String): Path = {
+    logger.debug(s"getAbsolutePath: $modelSource")
+    Paths.get(localFolder, modelSource)
+  }
 }
